Moved header to post-edit folder
