--- packages/edit-site/src/components/post-edit-header/style.scss
@@ -1,9 +1,0 @@
-.edit-site-post-edit-header {
-	.edit-site-post-edit-header__description {
-		color: $gray-700;
-	}
-
-	.edit-site-post-edit-header__title {
-		@include heading-medium();
-	}
-}
