+++ packages/edit-site/src/components/post-edit/style.scss
@@ -14,3 +14,13 @@
 		padding-top: $grid-unit-20;
 	}
 }
+
+.edit-site-post-edit-header {
+	.edit-site-post-edit-header__description {
+		color: $gray-700;
+	}
+
+	.edit-site-post-edit-header__title {
+		@include heading-medium();
+	}
+}
